requirements.txt
boto3==1.42.3
pwinput==1.0.3
python-dotenv==1.2.1
segno==1.6.6
